Initial upload of IO handlers
Contains the structure for ioFile onject. All file handling can be done with class.
Has issues when trying to write to a class using pointers. Had to implement temporary solution in which file is opened for every write then closed again.
Slow for writing for large files, TODO implement pointer for writing to files.